Made massive changes to CatKernel: Added virtual memory manager, changed task list from linked list to queue, fixed massive virtual memory mapping bug, and added TinyUBSan

+++ src/lib/common/memcmp.c
@@ -4,7 +4,7 @@ int memcmp(const void * addr1, const void * addr2, size_t count) {
   char * m1 = (char *)addr1;
   char * m2 = (char *)addr2;
   size_t ret = 0;
-  for(int i = 0; i < count; i++)
+  for(size_t i = 0; i < count; i++)
     ret += (m1[i] != m2[i]);
   return ret;
 }
